Modifies delta-chi-squared logic for fit to data
This changes nothing for a sensitivity study, because the minimum of
the chi-squared curve should always be at zero signal. However for a
fit to data, the minimum is often not at zero. When calculating
delta-chi-squared, we should only subtract chi-squared-zero if it is
less than (not greater than) the minimum of the curve.

Also adds a cool progress bar to `Limit`, using `tqdm` and minimises
some repeated code in generating the logging output.

Fixes some bugs with `group_name` in `store`.

Modifies `klz_majoron_limits` script accordingly.

Also fixes a bug in `LimitResults` and a bit of tidying up in `Fit`.